LayoutTests/storage/indexeddb/IDBTransaction-page-cache-expected.txt

Tests that an IDBRequest & IDBIndex does not prevent page caching.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from page cache
PASS !!myIndex is true
pagehide - entering page cache
pageshow - from page cache
PASS Page entered page cache
PASS successfullyParsed is true

TEST COMPLETE
